maj 10/03

+++ backEnd/controllers/deleteType.js
@@ -1,6 +1,7 @@
 import Comment from "../models/comSchema.js";
 import Music from "../models/musicSchema.js";
 import User from "../models/userSchema.js";
+import { getUserIdFromToken } from "../utils/utils.js";
 import fs from "fs"
 
 export const deleteType = async (req,res) => {
@@ -13,31 +14,39 @@ export const deleteType = async (req,res) => {
         case 'comment': type = Comment; break
     }
     
-    const id = req.params.id
+    const id = req.params.id;
 
     try {
-        const item = await type.findByIdAndDelete(id);
-
+        const userId = getUserIdFromToken(req);
+        const item = await type.findById(id)
+        
         if(!item){ 
-            res.status(400).json({message: 'Element introuvable'})
-        }else{
-            fs.unlink(`public/${item.audio}`, (err) => {
-                if (err) {
-                    res.status(500).json({message: 'Erreur lors de la suppression du fichier audio'})
-                }
-                
-                if(!item.image){
-                    res.status(200).json({message: 'Suppression réussi'})
-                }else{
-                    fs.unlink(`public/${item.image}`, (err) => {
-                        if (err) {
-                            return res.status(500).json({message: 'Erreur lors de la suppression du fichier image'})
-                        }
-                        return res.status(200).json({message: 'Suppression réussi'})
-                    });
-                }
-            });
+            return res.status(400).json({message: 'Element introuvable'})
+        };
+
+        if(item.user.toString() !== userId){
+            return res.status(500).json({message: 'Vous ne pouvez pas supprimer ce fichier'})
         }
+
+        fs.unlink(`public/${item.audio}`, (err) => {
+            if (err) {
+                return res.status(500).json({message: 'Erreur lors de la suppression du fichier audio'})
+            }
+            
+            if(!item.image){
+                return res.status(200).json({message: 'Suppression réussi'})
+            }else{
+                fs.unlink(`public/${item.image}`, (err) => {
+                    if (err) {
+                        return res.status(500).json({message: 'Erreur lors de la suppression du fichier image'})
+                    }
+                    return res.status(200).json({message: 'Suppression réussi'})
+                });
+            }
+        });
+
+        await type.findByIdAndDelete(id);
+        
     } catch (err) {
         return res.status(400).json({message: 'Erreur'});
     }
